Including Bulk Authentication Delete. (#5)
This commit is to include a new request and response message for a bulk
authorization delete for a single resource. It includes a descriptions
and the configurations to send this request and the expected succesful
and failied responses.

A typo was also in the 'Securing an XMPP Multi-User chat' flow.